Create Polar_Angles
Given a list of points in the 2D plane, sort them in ascending order of their polar angle. In case multiple points share exactly the same polar angle, the one with lesser distance from the origin (0,0) should occur earlier in the sorted list.

Input Format 
The first line contains an integer N. 
This is followed by N lines containing pairs of space separated integers, x and y which represent the coordinates of the points in the cartesian plane.

Constraints 
1 <= N <= 1000 
-100 <= x,y <= 100 
The point (0,0) will not be present in the list of points.

Output Format 
The output should contain N lines. Each line should contain two integers x and y, corresponding to a point in the original list. Display the points in ascending order of their polar angle.